backend/requirements.txt
python-dotenv
fastapi
uvicorn
requests
pydantic
scikit-learn
joblib
